Don't add nav unless menu exists
Nice looking theme Zachary. Just came across this after looking for some ideas for a <a href="https://github.com/Wordpress-Development/genesis-bootstrap">plugin I was working on</a> that integrates bootstrap for the genesis theme.

Anyway, thought this might be better to check if the primary menu is there before adding the <nav> code. 

Keep up the good work! 

Bryan